feat(api): add support for HATEOAS JSON media type
This change adds support for the "application/vnd.amard-ecc.hateoas+json"
media type in the API. This allows clients to request HATEOAS-enabled
responses, which include links to related resources.

The changes include:

- Configuring the MVC options to add the custom media type to the
  supported media types for the NewtonsoftJsonOutputFormatter.
- Adding a new `CustomeMediaTypeNames` class to define the custom media
  type name.
- Updating the `AppSettingsController` to check the `Accept` header and
  include links in the response if the custom media type is requested.

These changes improve the discoverability and navigability of the API,
making it easier for clients to interact with the application.